requirements.txt
# Core Spark
pyspark==4.1.1
delta-spark==4.2.0

# Data handling
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0   # Arrow serialization, required for Spark DataFrame <-> Pandas conversion

# Visualization (optional, useful in dev/notebooks)
matplotlib==3.9.0
seaborn==0.13.2

# Jupyter (optional, for interactive development)
jupyterlab==4.2.3

# Cloud/Storage connectors (include only what you need)
boto3==1.34.0              # AWS S3 integration
azure-storage-blob==12.19.0 # Azure Blob Storage
google-cloud-storage==2.16.0 # GCP Storage

# Utilities
requests==2.32.3   # For API calls